+++ R/plotting.R
@@ -406,16 +406,11 @@ plot2D <- function(object,
     } else if (method == "PCA") {
         if (missing(methargs))
             methargs <- list(scale = TRUE, center = TRUE)
-        .pca <- do.call(prcomp, c(list(x = exprs(object)),
-                                  methargs))
-        .data <- .pca$x[, dims]
-        .vars <- (.pca$sdev)^2
-        .vars <- (.vars / sum(.vars))[dims]
-        .vars <- round(100 * .vars, 2)
-        .xlab <- paste0("PC", dims[1], " (", .vars[1], "%)")
-        .ylab <- paste0("PC", dims[2], " (", .vars[2], "%)")
-        colnames(.data) <- c(.xlab, .ylab)
-    } else if (method == "MDS")  { ## MDS
+        .data <- dimred(object, method = method, methargs = methargs)
+        .data <- .data[, dims]
+        .xlab <- colnames(.data)[1]
+        .ylab <- colnames(.data)[2]
+    } else if (method == "MDS")  { 
         if (!missing(methargs))
             warning("'methargs' ignored for MDS")
         ## TODO - use other distances